% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_plot.R
\name{cor_plot}
\alias{cor_plot}
\title{correlation plot}
\usage{
cor_plot(model, greg = TRUE, axis1 = NULL, axis2 = NULL, n_draws = 200, ...)
}
\arguments{
\item{model}{result of \code{\link{sociality_model}}}

\item{greg}{description}

\item{axis1}{name of behavior 1}

\item{axis2}{name of behavior 2}

\item{n_draws}{numeric, number of draws to be plotted}

\item{...}{additional argument to \code{\link{plot}}}
}
\value{
a plot
}
\description{
correlation plot
}
\examples{
model <- fit_example_model(num = 1)
cor_plot(model, xlim = c(-3, 3), asp = 1)
cor_plot(model, greg = FALSE, xlim = c(-3, 3), asp = 1)
}
