% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_multi.R
\name{generate_data_multi}
\alias{generate_data_multi}
\title{generate multiple groups with features from a underlying mechanisms}
\usage{
generate_data_multi(groups, indisds, dyadsds, bintercepts)
}
\arguments{
\item{groups}{list with group compositions (see \code{\link{generate_groups}})}

\item{indisds, dyadsds}{vector with SD}

\item{bintercepts}{vector with intercepts}
}
\value{
a list with matrices that can be passed to
\code{\link{make_stan_data_from_matrices_multi}}
}
\description{
please consider this still EXPERIMENTAL!
}
\details{
The function is a wrapper for \code{\link{generat_data}}.
}
\examples{
n <- 3 # number of groups to generate
# just assign group size and id codes
groups <- generate_groups(n_groups = n, min_nids = 4, max_nids = 5)
# set true parameter value
(indisds <- round(exp(rnorm(n, -1, 0.8)), 1))
(dyadsds <- round(exp(rnorm(n, -0.5, 0.6)), 1))
(bintercepts <- round(rnorm(n, -2, 1), 1))
generate_data_multi(groups, indisds, dyadsds, bintercepts)
}
