% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{generate_groups}
\alias{generate_groups}
\title{helper for creating ids over multiple groups}
\usage{
generate_groups(
  n_groups = 10,
  min_nids = 4,
  max_nids = 12,
  poolsize = (max_nids * 1.2)
)
}
\arguments{
\item{n_groups}{number of groups to be generated}

\item{min_nids, max_nids}{range of group sizes}

\item{poolsize}{integer total number of unique individuals across all groups}
}
\value{
a list
}
\description{
helper for creating ids over multiple groups
}
