% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{grooming}
\alias{grooming}
\title{grooming matrices and observation effort for four macaque species}
\format{
a list with four items. Each item is a list itself with one grooming
matrix and a vector or matrix for observation effort.
}
\source{
\itemize{
\item{\emph{Macaca assamensis}:} {Julia Ostner and Oliver Schülke provided data on Assamese macaques from Phu Khieo Wildlife Sanctuary.}
\item{\emph{M. fuscata}:} {Andrew MacIntosh}
\item{\emph{M. nigra}:} {Julie Duboscq}
\item{\emph{M. sylvanus}:} {Julia Fischer}
}
}
\usage{
grooming
}
\description{
grooming matrices and observation effort for four macaque species
}
\keyword{datasets}
