% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{make_indices}
\alias{make_indices}
\title{indexing over multiple groups or periods}
\usage{
make_indices(grouplist)
}
\arguments{
\item{grouplist}{list with character vectors}
}
\value{
list with two data frames
}
\description{
indexing over multiple groups or periods
}
