% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_data_from_matrices_multi.R
\name{make_stan_data_from_matrices_multi}
\alias{make_stan_data_from_matrices_multi}
\title{prepare data for Stan for multi-group model}
\usage{
make_stan_data_from_matrices_multi(
  bmats,
  btypes = c("count", "prop"),
  omats = NULL,
  forsimeval = FALSE,
  indidata = NULL,
  dyaddata = NULL
)
}
\arguments{
\item{bmats}{list with named(!) interaction matrices (undirected,
upper triangle only)}

\item{btypes}{character: either \code{"count"} or \code{"prop"}}

\item{omats}{optional list of named matrices with observation effort}

\item{forsimeval}{internal, for simulations only}

\item{indidata}{internal, for simulations only}

\item{dyaddata}{internal, for simulations only}
}
\value{
a list ready to be passed to Stan
}
\description{
prepare data for Stan for multi-group model
}
\examples{
data("grooming")
bmats <- list(nig = grooming$nig$groom, ass = grooming$ass$groom)
omats <- list(nig = grooming$nig$obseff, ass = grooming$ass$obseff)
bamoso:::make_stan_data_from_matrices_multi(bmats, "count", omats)
}
