% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_model_dens.R
\name{pp_model_dens}
\alias{pp_model_dens}
\title{Posterior predictive checks for continuous or discrete data (via density)}
\usage{
pp_model_dens(
  mod_res,
  xvar = 1,
  n_draws = 20,
  plot_draws = TRUE,
  xlim = NULL,
  ylim = NULL,
  xadjust = NULL,
  selected_id = NULL,
  print_info = FALSE,
  ...
)
}
\arguments{
\item{mod_res}{model result from \code{\link{sociality_model}}}

\item{xvar}{numeric or character, the index or name of the behavior to
be plotted. Default is \code{1}, i.e. the first behavior.}

\item{n_draws}{numeric}

\item{plot_draws}{logical, should the random draws actually be plotted
(default is \code{TRUE})}

\item{xlim, ylim}{numeric, each of length two. Overrides axis limits gleaned
from the data. Default is \code{NULL} (take values from the
model and data)}

\item{xadjust}{numeric}

\item{selected_id}{character of length 1, if provided: plot only data for
that individual. If left at its default (\code{NULL}),
plot aggregate over all individuals.}

\item{print_info}{logical, default is \code{FALSE}. If \code{TRUE}, some
summaries are printed to screen for the selected
individual. (This argument is ignored if
\code{selected_id = NULL}).}

\item{...}{further arguments for \code{\link{density}}}
}
\value{
a plot
}
\description{
Posterior predictive checks for continuous or discrete data (via density)
}
