% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{prob2lin}
\alias{prob2lin}
\title{link function for probabilities with observation effort}
\usage{
prob2lin(x, obseff)
}
\arguments{
\item{x}{probability predictor (constrained)}

\item{obseff}{positive observation effort (offset)}
}
\description{
link function for probabilities with observation effort
}
