% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_foo.R
\name{sim_datasets}
\alias{sim_datasets}
\title{create simulated data sets}
\usage{
sim_datasets(
  n_ids = sample(5:60, 1),
  indi_sd = runif(1, 0.1, 3),
  dyad_sd = runif(1, 0.1, 3),
  n_beh = sample(3, 1),
  indi_cor = NULL,
  dyad_cor = NULL,
  behav_types = NULL,
  intercepts = runif(n_beh, -3, 1),
  disp_pars_gamma = runif(n_beh, 1, 20),
  disp_pars_beta = runif(n_beh, 5, 30),
  outpath = NULL,
  outfilename = NULL,
  silent = TRUE
)
}
\arguments{
\item{n_ids}{see \code{\link{generate_data}}}

\item{indi_sd}{numeric, SD for gregariousness (defaults to
\code{runif(1, 0.1, 3)})}

\item{dyad_sd}{numeric, SD for dyadic affinity (defaults to
\code{runif(1, 0.1, 3)})}

\item{n_beh}{see \code{\link{generate_data}}}

\item{indi_cor, dyad_cor}{numeric value for correlation parameter
(only works if \code{n_beh = 2}). Default is \code{NULL}, i.e.,
no correlations.}

\item{behav_types}{see \code{\link{generate_data}}}

\item{intercepts}{numeric of length two, determines the range for
intercepts. Defaults to \code{c(-3, 1)}. How many
intercepts are needed is determined inside the
function.}

\item{disp_pars_gamma}{see \code{\link{generate_data}}}

\item{disp_pars_beta}{see \code{\link{generate_data}}}

\item{outpath}{character, output path (defaults to R's \code{\link{tempdir}})}

\item{outfilename}{character, output file name (defaults to a random string)}

\item{silent}{logical, suppress all textual output}
}
\value{
a list
}
\description{
create simulated data sets
}
