% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sociality_plot.R
\name{sociality_plot}
\alias{sociality_plot}
\title{plot posteriors of individual and dyadic sociality SDs}
\usage{
sociality_plot(
  mod_res,
  xlim = c(0, 4),
  labs = list(indi = "gregariousness", dyad = "affinity"),
  do_legend = TRUE,
  do_dyadic = TRUE,
  do_indi = TRUE
)
}
\arguments{
\item{mod_res}{model result from \code{\link{sociality_model}}}

\item{xlim}{limits for horizontal axis}

\item{labs}{named list with labels for the individual- and
dyad-specific component}

\item{do_legend}{logical, should the legend be plotted
(default is \code{TRUE})}

\item{do_dyadic}{logical, should the posterior for the dyad-level
estimate be plotted (default is \code{TRUE})}

\item{do_indi}{logical, should the posterior for the individual-level
estimate be plotted (default is \code{TRUE})}
}
\value{
a plot
}
\description{
plot posteriors of individual and dyadic sociality SDs
}
