% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_data_from_association.R
\name{make_stan_data_from_association}
\alias{make_stan_data_from_association}
\title{convert group-by-individual association data into data list ready for Stan}
\usage{
make_stan_data_from_association(asso_table)
}
\arguments{
\item{asso_table}{a 0/1 matrix where columns correspond to individuals and
rows to individual samples}
}
\value{
a list
}
\description{
convert group-by-individual association data into data list ready for Stan
}
\details{
For now differences in sampling periods are ignored: each
individual is assumed to be present (theoretically observable)
in each sample.

\if{html}{\out{<div class="sourceCode">}}\preformatted{     If supplied via column names in \code{asso_table}, id codes are
     present in the output as names of vector of zeros.
}\if{html}{\out{</div>}}
}
\examples{
# generate some data
x <- sim_data_whitehead2015(n_ind = 5, n_periods = 9, ignore_temporal = TRUE)
assos <- x$assolist
head(assos)

make_stan_data_from_association(assos)

}
